+++ hltbwearos/src/main/AndroidManifest.xml
@@ -3,13 +3,22 @@
 
     <uses-permission android:name="android.permission.INTERNET"/>
 
+    <uses-feature
+        android:name="android.software.leanback"
+        android:required="true" />
+
+    <uses-feature
+        android:name="android.hardware.touchscreen"
+        android:required="false" />
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.HLTBwearos">
+        android:theme="@style/Theme.HLTBwearos"
+        android:banner="@mipmap/ic_launcher_round">
         <activity
             android:name="com.croumy.hltbwearos.MainActivity"
             android:exported="true"
@@ -18,7 +27,7 @@
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
+                <category android:name="android.intent.category.LEANBACK_LAUNCHER"/>
             </intent-filter>
 
             <intent-filter android:autoVerify="true">
